Removing tests of index.jsx
The tests are currently failing, as they cannot import
logger from nrfconnect/core. Mocking it would require adding an
extra file that contains the mock implementation, plus configuring
this in package.json. We want the boilerplate to be simple, so
I suggest removing these tests. I think these tests are not
valuable enough to justify complicating the boilerplate with mocks.

Keeping the unit test of the Hello component though, as it is
good to have a simple example of unit testing.