Fix the info messages on switching panes
Previously the pane components were only rendered when they are also
visible. But since shared 4.8.5 (which is used in the launcher since
3.5.0) all panes are rendered all the time.

This led to an error: The message "Showing Hello pane" was shown only
once, when starting the app and "Hiding Hello pane" was shown never.

With the new `active` prop, this can be corrected.